pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "langchain>=1.0.2",
    "langchain-community>=0.3.0",
    "langchain-ollama>=1.0.0",
    "langchain-tavily>=0.2.12",
    "langchainhub>=0.1.14",
    "langgraph>=0.2.0",
    "python-dotenv>=1.2.1",
]
